ProofBase

Description

This module contains operators for annotating equation reasonings and basic functions coppied from Haskell report.

Documentation

data SideExprInfo

Constructor for expression info, QED and Postulate.

Constructors

L ExprInfo

R ExprInfo

QED

Postulate

data ExprInfo

Constructor for expression info

Constructors

Def String

Prop String

Inst String

Beta

Eta

DefRec String Integer

(--.) :: a -> SideExprInfo -> a

Operator for adding annotation to expression

(===) :: a -> a -> a infixl 0

Main operator for checking type equality and combining expressions into chain.

(.) :: (b -> c) -> (a -> b) -> a -> c infixr 9

($) :: (a -> b) -> a -> b infixr 0

id :: a -> a

flip :: (a -> b -> c) -> b -> a -> c

const :: a -> b -> a